feat: rename tokens to cookies, update cookies
